feat: add icon button and small refactor
